Replace solve with create_solver_factory; refactor code
