+++ optapy-quickstarts/school-timetabling/main.py
@@ -1,6 +1,6 @@
 from functools import reduce
 
-from optapy import get_class, solve
+from optapy import get_class, solver_factory_create
 from optapy.types import SolverConfig, Duration
 from domain import TimeTable, Lesson, generate_problem
 from constraints import define_constraints
@@ -75,6 +75,8 @@ solver_config = SolverConfig().withEntityClasses(get_class(Lesson)) \
     .withConstraintProviderClass(get_class(define_constraints)) \
     .withTerminationSpentLimit(Duration.ofSeconds(30))
 
-solution = solve(solver_config, generate_problem())
+solver = solver_factory_create(solver_config).buildSolver()
+
+solution = solver.solve(generate_problem())
 
 print_timetable(solution)
